requirements.txt
biopython >= 1.83
numpy >= 1.26.4
pandas >= 2.2.2
